driver: regulator: Silence the pointer to u8 size warning
The pointer returned by of_device_get_match_data() didn't have the
same size as u8 on 64-bit architectures, causing a compile warning.
Cast the return value to unsigned long and then to u8 to silence
the warning.

+++ drivers/regulator/qpnp-lcdb-regulator.c
@@ -2230,8 +2230,7 @@ static int qpnp_lcdb_regulator_probe(struct platform_device *pdev)
 		return -EINVAL;
 	}
 
-	lcdb->subtype = (u8)of_device_get_match_data(&pdev->dev);
-
+	lcdb->subtype = (u8)(unsigned long)of_device_get_match_data(&pdev->dev);
 	lcdb->dev = &pdev->dev;
 	lcdb->pdev = pdev;
 	mutex_init(&lcdb->lcdb_mutex);
